+++ src/BasicInputs/index.ts
@@ -1,5 +1,5 @@
 export * from './BooleanInput';
-export * from './DatetimeInput';
+export * from './DatetimeInput'
 export * from './NumberInput';
 export * from './SelectInput';
 export * from './StringInput';
