+++ src/BasicInputs/stories/DatetimeInput.stories.tsx
@@ -1,14 +1,14 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { DatetimeInput2 } from '../DatetimeInput2';
+import { DatetimeInput } from '../DatetimeInput';
 
 const meta = {
   title: 'Commons/DateTime',
   tags: ['autodocs'],
-  component: DatetimeInput2,
+  component: DatetimeInput,
   args: {
     value: '2024-07-10T01:00:00'
   }
-} satisfies Meta<typeof DatetimeInput2>;
+} satisfies Meta<typeof DatetimeInput>;
 export default meta;
 
 type Story = StoryObj<typeof meta>;
